% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_parameters}
\alias{default_parameters}
\title{2015 MEND plus parameters}
\format{A data table consisting of 4 columns with inputs from Wang et al. 2013, doi:10.1890/12-0681.1.
\describe{
\item{parameter}{String character of the default MEND parameters.}
\item{description}{String character describing the parameter.}
\item{units}{String character of the parameter units.}
\item{value}{Numeric values taken from the table 2 of the Wang et al. 2013}
}}
\usage{
default_parameters
}
\description{
The default parameters for MEND 2015.
}
\concept{parameters}
\keyword{datasets}
