% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{log_prior}
\alias{log_prior}
\title{Calculate the log prior}
\usage{
log_prior(params, ptable = default_parameters)
}
\arguments{
\item{params}{a vector of parameters, the vector must include parameter names.}

\item{ptable}{a MEND parameter table that will be updated with the values from params. \link{default_parameters}
is used as the default paraemter table. However ptable may be set to any data frame that contains the
following columns.
\describe{
\item{parameter}{String character of the default MEND parameters.}
\item{description}{String character describing the parameter.}
\item{units}{String character of the parameter units.}
\item{value}{Numeric values taken from the table 2 of the Wang et al. 2013}}}
}
\value{
the log prior for the sampled parameters.
}
\description{
\code{log_prior} Right now this is set up to calculate the log prior for a set of
parameters based on the normal distribtuion.
TODO there should probably be some better way to set up and define the hyperparameters.
}
