% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{make_logpost}
\alias{make_logpost}
\title{Set up the function that will calculate the log posterior.}
\usage{
make_logpost(
  comp,
  params,
  t = seq(0, 1000, 1),
  state = c(P = 10, M = 5, Q = 0.1, B = 2, D = 1, EP = 1e-05, EM = 1e-05, IC = 0, Tot =
    18.10002),
  carbon_pools_func = MEND_carbon_pools,
  flux_func = MEND_fluxes,
  ptable = default_parameters,
  verbose = FALSE
)
}
\arguments{
\item{comp}{\describe{the comparison data frame it must comain at least two columns,
\item{time}{model time}
\item{IC}{inorganic carbon, is the the co2 flux from  microbial respiration.}}}

\item{params}{a named vector of the inital guess of the parameters that will be calibrated in the mcmc.}

\item{t}{time vector that will be used to run the model.}

\item{state}{the inital state variables}

\item{carbon_pools_func}{the function that governs the relationship between the different carbon
pools, by default it is set to \code{MEND_carbon_pools}.}

\item{flux_func}{the function that governs fluxes, by default is it set to \code{MEND_fluxes}.}

\item{ptable}{\describe{ A table of MEND parameter calues, by default is it set to \cpde{default_parameters}. While
a different table may be used it must contain the following columns.
\item{parameter}{String character of the default MEND parameters.}
\item{description}{String character describing the parameter.}
\item{units}{String character of the parameter units.}
\item{value}{Numeric values taken from the table 2 of the Wang et al. 2013}}}

\item{verbose}{a TRUE/FALSE indicator to run the function in verbose mode where messages are printed out,
by default set to FALSE.}
}
\value{
a function that will return the log posterior.
}
\description{
\code{make_logpost} Define the function that will calculate the log
posterior. The function returned by \code{make_logpost} will be used in
the \code{metrop} to conduct the mcmc.
TODO figure out  a better way to document the inputs/
}
