% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEND2013.R
\name{MEND2013_fluxes}
\alias{MEND2013_fluxes}
\title{Define the 2013 MEND fluxes.}
\usage{
MEND2013_fluxes(state, parms)
}
\arguments{
\item{state}{A numeric vector of the different MEND carbon pool states.}

\item{parms}{A data frame of the parameters.}
}
\value{
A list of functions that calculate the fluxes between \href{https://doi.org/10.1890/12-0681.1}{Wang et al. 2013} carbon pools.
}
\description{
\code{MEND2013_fluxes} Defines a system of equations that
describe the state of the fluxes between the pools from \href{https://doi.org/10.1890/12-0681.1}{Wang et al. 2013}.
By retruning a list of named functions.
}
\seealso{
Other 2013 MEND model functions: 
\code{\link{MEND2013_pools}()}
}
\concept{2013 MEND model functions}
\concept{carbon flux functions}
