% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MEND2013_params}
\alias{MEND2013_params}
\title{The default MEND 2013 parameters.}
\format{A data table consisting of 4 columns with inputs from \href{https://doi.org/10.1890/12-0681.1}{Wang et al. 2013}.
\describe{
\item{parameter}{String character of the default MEND parameters.}
\item{description}{String character describing the parameter.}
\item{units}{String character of the parameter units.}
\item{value}{Numeric values taken from the table 2 of the Wang et al. 2013}
}}
\usage{
MEND2013_params
}
\description{
The default MEND 2013 parameters.
}
\concept{parameters}
\keyword{datasets}
