% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEND2013.R
\name{MEND2013_pools}
\alias{MEND2013_pools}
\title{Define 2013 MEND carbon pools}
\usage{
MEND2013_pools(t, state, parms, flux_function = MENDplus::MEND2013_fluxes)
}
\arguments{
\item{t}{is for time}

\item{state}{A numeric vector of the different MEND carbon pool states.}

\item{parms}{A data frame of the parameters.}

\item{flux_function}{a function that will return a list of functions that modify how carbon moves between
the different MEND carbon pools the default is setup to reproduce the structure documented in
\href{https://doi.org/10.1890/12-0681.1}{Wang et al. 2013}.}
}
\value{
A list of the state variables
}
\description{
\code{MEND2013_pools} Defines the system of equations that
describe the state of the carbon pools from \href{https://doi.org/10.1890/12-0681.1}{Wang et al. 2013}.
}
\references{
\href{https://doi.org/10.1890/12-0681.1}{Wang et al. 2013}
}
\seealso{
Other 2013 MEND model functions: 
\code{\link{MEND2013_fluxes}()}
}
\concept{2013 MEND model functions}
\concept{carbon pool functions}
