% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEND2pools.R
\name{MEND2pool_fluxes}
\alias{MEND2pool_fluxes}
\title{Define carbon fluxes for when there are 2 carbon pools}
\usage{
MEND2pool_fluxes(state, parms)
}
\arguments{
\item{state}{A numeric vector of the different MEND carbon pool states}

\item{parms}{A data frame of the parameters, default is set to \code{MEND2pool_params}}
}
\value{
A list of the state variables
}
\description{
\code{MEND2pool_fluxes} Defines a set equations that determine the fluxes between the \code{MEND2pool_pools}
describe the state of the carbon pools modified from  \href{https://doi.org/10.1890/12-0681.1}{Wang et al. 2013}.
}
\seealso{
Other 2013 MEND 2 pool model: 
\code{\link{MEND2pool_pools}()}

Other carbon flux functions: 
\code{\link{MEND2013_fluxes}()}
}
\concept{2013 MEND 2 pool model}
\concept{carbon flux functions}
