% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEND2pools.R
\name{MEND2pool_pools}
\alias{MEND2pool_pools}
\title{Define carbon pools for when there are 2 carbon pools}
\usage{
MEND2pool_pools(
  t,
  state,
  parms = MEND2pool_params,
  flux_function = MEND2pool_fluxes
)
}
\arguments{
\item{t}{is for time}

\item{state}{A numeric vector of the different MEND carbon pool states}

\item{parms}{A data frame of the parameters, default is set to \code{MEND2pool_params}}

\item{flux_function}{a function that will return a list of functions that modify how carbon moves between
the different MEND carbon pools default is set to \code{MEND2pool_fluxes}}
}
\value{
A list of the state variables
}
\description{
\code{MEND2pool_pools} Defines the system of equations that
describe the state of the carbon pools modified from  \href{https://doi.org/10.1890/12-0681.1}{Wang et al. 2013}.
}
\seealso{
Other 2013 MEND 2 pool model: 
\code{\link{MEND2pool_fluxes}()}

Other carbon pool functions: 
\code{\link{MEND2013_pools}()}
}
\concept{2013 MEND 2 pool model}
\concept{carbon pool functions}
