% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEND_Cpools.R
\name{MEND_carbon_pools}
\alias{MEND_carbon_pools}
\title{Carbon pools}
\usage{
MEND_carbon_pools(t, state, parms, flux_function = MEND_fluxes)
}
\arguments{
\item{t}{is for time}

\item{state}{A numeric vector of the different MEND carbon pool states.}

\item{parms}{A data frame of the parameters.}

\item{flux_function}{a function that will return a list of functions that modify how carbon moves between
the different MEND carbon pools the default is set to \code{MEND_fluxes} to reproduce the 2013 Wang et. al
paper results. This function must beable to use the same state and parms as inputs.}
}
\value{
A list of the state variables.
}
\description{
\code{MEND_carbon_pools} Define a system of equations that
describe the state of the carbon pools from Wang et al. 2013.
}
