% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEND_fluxes.R
\name{MEND_fluxes}
\alias{MEND_fluxes}
\title{Define the stanadard MEND fluxes.}
\usage{
MEND_fluxes(state, parms)
}
\arguments{
\item{state}{A numeric vector of the different MEND carbon pool states.}

\item{parms}{A data frame of the parameters.}
}
\value{
A list of functions that define
}
\description{
\code{MEND_fluxes} Define a system of equations that
describe the state of the fluxes between the pools from Wang et al. 2013.
By retruning a list of named functions. We will be able to modify how
the fluxes are defined and therefore used in \code{MEND_carbon_pools}.
}
