% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{modify_fluxes_func}
\alias{modify_fluxes_func}
\title{Alter a MEND function}
\usage{
modify_fluxes_func(params, state, flux_func, replace_with)
}
\arguments{
\item{params}{data.table containing the following columns: parameter, value, and units}

\item{state}{a vector of the intial state values, must be named}

\item{flux_func}{a MEND flux object such as the \code{MEND2013_fluxes}}

\item{replace_with}{a list containing the new elements of the functions}
}
\description{
Helper function that can be used to modify \code{MEND2013_fluxes}.
To help test out different structural configurations of MEND.
}
\seealso{
Other helper function: 
\code{\link{solver}()}
}
\concept{helper function}
