% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{replace_params}
\alias{replace_params}
\title{Change parameter values in a parameter table.}
\usage{
replace_params(params, ptable = MENDplus::default_parameters)
}
\arguments{
\item{params}{a vector of parameters, the vector must include parameter names.}

\item{ptable}{a MEND parameter table that will be updated with the values from params. \link{default_parameters}
is used as the default paraemter table. However ptable may be set to any data frame that contains the
following columns.
\describe{
\item{parameter}{String character of the default MEND parameters.}
\item{description}{String character describing the parameter.}
\item{units}{String character of the parameter units.}
\item{value}{Numeric values taken from the table 2 of the Wang et al. 2013}}}
}
\value{
A MEND parameter table with updated params values.
}
\description{
\code{replace_params} Update a table of MEND parameter values such as \link{default_parameters},
with new parameter values. This a helper function to run the mcmc and can also be helpful in
a sensitivity analysis
}
